Bruk eøs-toggle for å rendere eøs-routes
