+++ src/frontend/App.tsx
@@ -11,6 +11,7 @@ import { EøsProvider } from './context/EøsContext';
 import { FeatureTogglesProvider } from './context/FeatureToggleContext';
 import { InnloggetProvider } from './context/InnloggetContext';
 import { LastRessurserProvider } from './context/LastRessurserContext';
+import { RoutesProvider } from './context/RoutesContext';
 import { StegProvider } from './context/StegContext';
 import { routerBasePath } from './Miljø';
 import { GlobalStyle } from './Theme';
@@ -22,22 +23,24 @@ function App() {
                 <FeatureTogglesProvider>
                     <AppProvider>
                         <EøsProvider>
-                            <Router basename={routerBasePath}>
-                                <StegProvider>
-                                    <GlobalStyle />
-                                    {process.env.NODE_ENV !== 'production' && (
-                                        <AlertStripe type="advarsel">
-                                            {`Denne siden er under utvikling. `}
-                                            <a href="https://www.nav.no/no/person/familie/barnetrygd-og-kontantstotte/barnetrygd">
-                                                Klikk her for å gå til våre sider for barnetrygd
-                                            </a>
-                                        </AlertStripe>
-                                    )}
-                                    <AppNavigationProvider>
-                                        <AppContainer />
-                                    </AppNavigationProvider>
-                                </StegProvider>
-                            </Router>
+                            <RoutesProvider>
+                                <Router basename={routerBasePath}>
+                                    <StegProvider>
+                                        <GlobalStyle />
+                                        {process.env.NODE_ENV !== 'production' && (
+                                            <AlertStripe type="advarsel">
+                                                {`Denne siden er under utvikling. `}
+                                                <a href="https://www.nav.no/no/person/familie/barnetrygd-og-kontantstotte/barnetrygd">
+                                                    Klikk her for å gå til våre sider for barnetrygd
+                                                </a>
+                                            </AlertStripe>
+                                        )}
+                                        <AppNavigationProvider>
+                                            <AppContainer />
+                                        </AppNavigationProvider>
+                                    </StegProvider>
+                                </Router>
+                            </RoutesProvider>
                         </EøsProvider>
                     </AppProvider>
                 </FeatureTogglesProvider>
