+++ src/frontend/context/RoutesContext.tsx
@@ -1,0 +1,68 @@
+import createUseContext from 'constate';
+
+import { EFeatureToggle } from '../typer/feature-toggles';
+import { IRoute, RouteEnum } from '../typer/routes';
+import { useFeatureToggles } from './FeatureToggleContext';
+
+export const omBarnetBasePath = 'om-barnet';
+export const eøsBarnBasePath = 'eøs-barn';
+
+const [RoutesProvider, useRoutes] = createUseContext(() => {
+    const { toggles } = useFeatureToggles();
+
+    const routes: IRoute[] = [
+        { path: '/', label: 'Forside', route: RouteEnum.Forside },
+        { path: '/om-deg', label: 'Om deg', route: RouteEnum.OmDeg },
+        {
+            path: '/din-livssituasjon',
+            label: 'Din Livssituasjon',
+            route: RouteEnum.DinLivssituasjon,
+        },
+        { path: '/velg-barn', label: 'Velg barn', route: RouteEnum.VelgBarn },
+        { path: '/om-barna', label: 'Om barna', route: RouteEnum.OmBarna },
+        {
+            path: `/${omBarnetBasePath}/barn-:number`,
+            label: `Om barnet`,
+            route: RouteEnum.OmBarnet,
+        },
+        ...(toggles[EFeatureToggle.EØS_KOMPLETT]
+            ? [
+                  {
+                      path: '/eøs-søker',
+                      label: 'Eøs søker',
+                      route: RouteEnum.EøsForSøker,
+                  },
+              ]
+            : []),
+        ...(toggles[EFeatureToggle.EØS_KOMPLETT]
+            ? [
+                  {
+                      path: `/${eøsBarnBasePath}/barn-:number`,
+                      label: `Om EØS barn`,
+                      route: RouteEnum.EøsForBarn,
+                  },
+              ]
+            : []),
+        {
+            path: '/oppsummering',
+            label: 'Oppsummering',
+            route: RouteEnum.Oppsummering,
+        },
+        {
+            path: '/dokumentasjon',
+            label: 'Dokumentasjon',
+            route: RouteEnum.Dokumentasjon,
+        },
+        {
+            path: '/kvittering',
+            label: 'Kvittering',
+            route: RouteEnum.Kvittering,
+        },
+    ];
+    const hentRouteObjektForRouteEnum = (routeEnum: RouteEnum) =>
+        routes.find(route => route.route === routeEnum) ?? routes[0];
+
+    return { routes, hentRouteObjektForRouteEnum };
+});
+
+export { RoutesProvider, useRoutes };
